Removed the "panguin" directory and added JeffersonLab/panguin as a submodule.
The "panguin/macros" directory was renamed to be "panguinMacros".
These macros are largely remants from PREX/CREX, and should be cleaned up and adapted for MOLLER.
Argueably, the panguin and root scripts could go in a separate repo and keep just the engine code within japan-MOLLER.

panguin is now a submodule and will be the head of JeffersonLab/panguin.