Added a new set of mock-data parameter files for the Hall A beamline, and increased the maximum size of the ROOT branch vector.
The mock_moller_beamline* map files have the device names and locations
for the Hall A beamline elements based on MOLLER DocDB 1071-v5.

This increased the number of variables in the root files, and we needed
to increase the maximum size of the ROOT branch vector to avoid run-time
errors.

The standard mock data detector map and the mock data parameter file
have been changed to match.