Fix array out of bounds error in QwHelicity::SetFirstBits (#70)
* Fix array out of bounds error in QwHelicity::SetFirstBits

SetFirstBits calls GetRandomSeed, which (a) only works with an input array of size 25 (24 words actually used) and (b) uses Fortran indexing into the input array (1-24). The original code apparently mixed up C and Fortran indexing, causing an out of bounds array access in GetRandomSeed. This is fixed here by shifting the input array indexing by one, as was probably intended. Additionally, an exception in thrown if the number of bits requested is not equal to 24, since 24 is the only value currently supported. To lift this limit, GetRandomSeed would need to be modified.

SetFirstBits is only called from exactly one place in this package, viz. QwMockDataGenerator.cc:194, with nbits=24, so the exception should never trigger. However, the results could be different since the original buggy code copied the bits of the seed argument into the wrong elements of the array.

[[A later PR will change from a 24-bit to 30-bit initialization routine.]]

* Update QwHelicity.cc

Fix clang error detected by CI

* Changed the initialization inside "SetFirstBits" to initialize the 25
bits for the mock data generator in the same way as we have been, but
with a total array size of nbits+1 instead of just nbits.

---------

Co-authored-by: Paul King <pking@jlab.org>